requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cssselect==1.1.0
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
fake-useragent==0.1.11
html5lib==1.0.1
http-requests==0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
multidict==4.7.3
oauthlib==3.1.0
parse==1.14.0
pdfdocument==3.3
Pillow==7.0.0
pyee==6.0.0
pylint==2.4.4
PyPDF2==1.26.0
pyppeteer==0.0.25
pyquery==1.4.1
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tqdm==4.41.1
urllib3==1.25.7
w3lib==1.21.0
webencodings==0.5.1
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
